Add jemalloc as a dependency on osx
Summary:
Building the code on mac failed with a message that Jemalloc could not be found. After running 'brew install jemalloc' the build succeeded.
This diff adds jemalloc to the brew command in build-osx.md

Test Plan:
This diff only includes changes to build instructions.
To verify correctness of the change one could try building the code with and without jemalloc installed on their machine.

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6702