[CI] Set build timeout in the configuration file
Summary:
This allow for changing the timeout from the configuration file. If not
set, the default value is 1h. It has the added advantage to not take
into account the docker base image build time.

Test Plan:
Set some short value and run the build. Check the process is killed as
expected after the timeout expired.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6715