[CI] Add a templating system to the configuration
Summary:
This is not used for now but will become very useful when the artifact
are managed by the configuration file, allowing to add an artifact to
several configurations without editing them all.

Depends on D6732.

Test Plan:
Run any CI build.
Update the configuratio file to use a template and check the feature
works as expected.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6733