[CI] Manage the build artifacts from the configuration
Summary:
This add an "artifacts" dictionary to the configuration, which maps a
glob pattern (relative to the build directory) to an artifact name
(relative to the artifact directory).

The artifact list is not guaranteed to be 100% correct, but it can't
make it worst.

Depends on D6733.

Test Plan: Run any CI build plan and look at the artifacts in Teamcity.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6735