[CI] Refactor the build by making it a class
Summary:
This is a refactor that makes it easier to follow the build logic.
It also cleans up the output a bit when running as a user build (not on
teamcity).

Depends on D6749.

Test Plan:
Run any CI build as a user.
Check there is no longer any teamcity message in the output.

Look at the diff CI builds and check the teamcity messges are printed
and the artifacts uploaded.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6750