[CI] Fix wrong configuration when there is no template
Summary:
An extra indentation has been mistakenly added, that caused the
parameters to be correctly applied when a template is set but not if
there is none.
Another issue is that the `build-ibd-no-assumevalid-checkpoint`
configuration had no template (but should have one), which in turns
caused the default timeout of 1h to be used. This is not enough and
caused the build to fail.
This diff fixes both issues.

Test Plan: Run the `build-ibd-no-assumevalid-checkpoint` CI build plan.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6757