[avalanche] Add the ability to remove and rescore peers
Summary:
These are useful operations.

The PeerManager curently doesn't care about internal fragmentation. In practice, this isn't a big deal because algorithm used are at worst log, and at best constant time. This still needs to be fixed eventually, and fragmentation is being tracked in order to do so.

Depends on D6772

Test Plan:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6773