[CI] Run the check target instead of a list of subtargets
Summary:
This will cause test_bitcoin to run twice when the `run_test_bitcoin`
facility is used, but this should not cause any noticeable time penalty
since the tests run in parallel with `ninja check`. As a reference the
whole `ninja check` run takes about 5s on my machine.

The main advantage is that it prevents from lowering the coverage when a
test suite is renamed or added.

Test Plan: Run any CI build.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6763