Cleanup leftover phononactivationtime option
Summary:
Normally, post-upgrade we would replace `Params.phononActivationTime`
with `Params.phononHeight` similar to what we did with the Graviton upgrade.
However, doing so breaks the `abc-miner-fund.py` functional test.  There is some
value to keeping the that test maintained until it's clear we will not need it
in this coming upgrade feature freeze.  At that point, we can remove the test
and use phononHeight.

Cleaning up the unused runtime option gets rid of some kruft today and ensures
the patch for phononHeight will be minimal.

Test Plan: `ninja check check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6805