[CI] Improve teamcity error reporting by setting the failure message
Summary:
This turns the build status to failure with a dedicated message when a
timeout occurs, rather than exit code, and sets a better message in all
the other cases. It also makes it possible to search for the error in
the clean log file.

Test Plan: Run any CI build.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6817