Call RecoverDatabaseFile directly from wallettool
Summary:
When using the salvage command, call RecoverDatabaseFile directly
instead of SalvageWallet. Also removes SalvageWallet as it is no longer
needed.

SalvageWallet was doing an additional verify on the database which would
cause the salvage to sometimes fail. This is not needed.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19457 | core#19457]] [1/3]
https://github.com/bitcoin/bitcoin/pull/19457/commits/06e263a4e368671ebb4e4a77c1447ebd5104a488

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9639