[avalanche] Attach Proof to each Peer
Summary:
Right now, the proof simply consist of a random id and a score, but ultimately, we want these to be computed from data provided by the peers. By providing an opaque interface, we don't need to worry about this to move forward on the PeerManager.

Depends on D6847

Test Plan:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6848