doc: Use precise permission flags where possible
Summary:
Instead of mentioning the all-encompassing -whitelist* settings, change the docs to mention the exact permission flag that will influence the behaviour.

This is needed because in the future, the too-broad -whitelist* settings (they either include all permission flags or apply to all peers) might be deprecated to require the permission flags to be enumerated.

Alternatively, in the future there could be an RPC to set the net permission flags on an existing connection, in which case the -whitelist* terminology is of no help.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19474 | core#19474]]

Depends on D9831

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9832