Ignore cppcheck error in EraseOrphanTx()
Summary:
This is an unusual case where another container is actually tracking
the iterator of another container.  cppcheck `iterators1` is clearly useful,
so we do not want to disable it entirely. Instead, we can simply ignore this
one case.

Test Plan: `arc lint net_processing.cpp` is green after this patch

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6873