[net processing] Move tx & proof relay data to Peer
Summary:
Partial backport of [[https://github.com/bitcoin/bitcoin/pull/21160 | core#21160]]
https://github.com/bitcoin/bitcoin/pull/21160/commits/575bbd0dea6d12510fdf3220d0f0e47d969da6e9

Regarding the change in processor.cpp: `AddKnownProof` is now a static function in net_processing, so the call to add the local proof to the peer's known proofs has been moved there, just after `SendHello`.
The condition for calling this line of code is equivalent, the only possible change in behavior is that the proofid is no longer added to the known proofs if the schnorr signature fails in `Processor::sendHelloInternal` which would mean our sessionKey is invalid.

Depends on D14709

Test Plan: `ninja all check-all bitcoin-fuzzers`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14710