depends: enable lto support for Apple's ld64
Summary:
```
Note that this does not _enable_ lto by default in any way, only hooks
up the machinery for -flto to work correctly.

enable-lto-support is explicitly used for pinned-clang because we know
it works. It is neither enabled nor disabled in the external clang case
so that it can be auto-detected.
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/19530 | PR19530]].

Test Plan:
  ./contrib/teamcity/build-configurations.py build-osx

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D8784