[GITIAN] Remove vagrant support documentation for the gitian builds
Summary:
It is unmaintained and strongly outdated, so better remove it to avoid
confusion.
While updating `gitian-building.md` I also removed the table of content
which was an old version from core and not related to the current state
of the document.

Depends on D6976.

Test Plan: Read the doc.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6978