[GITIAN] Add documentation for building using Docker
Summary:
This is one of the alternatives methods offered by Gitian.
It is a bit simpler to setup, and does not require any user priviledge
to run which makes it a better candidate for our CI.

The VM-builder installation instructions have been removed because they
only apply to KVM machines which are not covered by our documentation.

Test Plan: Read and test the doc.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6989