Add visual accenting for the 'Create new receiving address' button
Summary:
> qt: Make "Create new receiving address" default unconditionally
>
>  This change makes this button visually accented, and gives to users a
> hint about the primary action.
> This commit does not change behavior.

> qt: Leverage the default "Create new receiving address" button
>
> When Enter or Return is pressed the default button will be always
> clicked. All buttons can always be clicked from the keyboard by pressing
> spacebar when the button has focus.

This is basically a simplification of the way pressing return in the
"Receive" tab of bitcoin-qt submits the form, doing it the standard Qt
way without additional code.

This is a backport of [[https://github.com/bitcoin-core/gui/pull/39 | core-gui#39]]

Test Plan:
`ninja && src/qt/bitcoin-qt`

In the Receive tab of the main window, verify that pressing Return still
submits the form and creates a new address.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10135