Rewrite parent txid loop of requested transactions
Summary:
Previously, we would potentially add the same txid many times to the rolling
bloom filter of recently announced transactions to a peer, if many outputs of
the same txid appeared as inputs in a transaction. Eliminate this problem and
avoid redundant lookups by asking the mempool for the unique parents of a
requested transaction.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19596 | core#19596]] [1/2]
https://github.com/bitcoin/bitcoin/pull/19596/commits/81961762439fb72fc2ef168164689ddc29d7ef94

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D10066