[CI] Enable Junit reporting via the build system and use it in scripts
Summary:
This diff adds a `--junit` option to the `build_cmake.sh` script, which
is a shortcut to enable junit reports. This option is set appropriately
by the build scripts as well as the ninja `-k0` flag that cause the
tests to continue in case of failure until dependencies can no longer be
met, so we gather as much result as possible in the report. This also
makes it possible to get rid of the hacky `run_test_bitcoin` wrapper
from the ci fixture.

The Junit reports are added to the artifacts and submitted to Teamcity.

Test Plan: Run the CI build plans.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7064