[GITIAN] Build with cmake 3.16 from the backport repository
Summary:
This diff adds a new `repositories` section to the YAML decription file
which allow to add a new repository to the sources list and install
packages from it.
This is used by our descriptors to install cmake 3.16 and use it for the
builds. This is a step toward enforcing cmake 3.16 for building the
project.

Test Plan:
I ran a gitian build locally with this version, and checked in the logs
that the expected cmake version is installed.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7072