[GITIAN] Stop distributing the *-unsigned.tar.gz archive
Summary:
This archive used to contain the installer for both win32 and win64 in a
single archive. Now that there is no win32 anymore it is just an archive
of the installer that offer no additional compression and can just be
removed.

Test Plan:
Run the windows Gitian build and check the archive is no longer part of
the output.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7089