[core#19724 6/9][refactor] Simplify check for block-relay-only connection.
Summary:
Previously we deduced it was a block-relay-only based on presence of the
m_tx_relay structure. Now we have the ability to identify it directly via a
connection type accessor function.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19724 | core#19724]] [6/9]
https://github.com/bitcoin/bitcoin/pull/19724/commits/1e563aed785565af6b7aed7f7399c52205d8f19c

Depends on D9774

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, majcosta, Fabien

Reviewed By: #bitcoin_abc, majcosta, Fabien

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9775