net: don't relay to the address' originator
Summary:
```
For each address to be relayed we "randomly" pick 2 nodes to send the address to (in `RelayAddress()`). However we do not take into consideration that it does not make sense to relay the address back to its originator (`CNode::PushAddress()` will do nothing in that case).

This means that if the originator is among the "randomly" picked nodes, then we will relay to one node less than intended.

Fix this by skipping the originating node when choosing candidates to relay to.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/19763 | core#19763]].

Depends on D10859.

Ref T1696.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D10860