[CMAKE] Make the split-debug script template executable
Summary:
This removes the need to create an intermediate script and copy it to
change the permissions. This is not perfect because the template should
really not be executed, but:
 - The copy prevent the file from being used as a dependency
 - The template is expected to move to a more appropriate template
directory which is less confusing to users.
This also makes sure it is properly tracked and regenerated if the
template changes.

Test Plan:
  cmake -GNinja ..
Check the split-debug.sh script is generated to the
`<build_dir>/contrib/devtools` directory and is executable (no change in
behavior).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7236