[GITIAN] Use the new build targets to split debug info and stripped bin
Summary:
Also cleans the install for osx as we now use cmake 3.16.
This makes the install part of the gitian scripts easier to understand
and maintain.
Since the default install target is used (in fact its -debug
counterpart), the man pages will now be part of the release shipment as
a positive side effect.

Depends on D7240, D7242 and D7243.

Test Plan:
Ran all the Gitian builds, checked the output are complete (now
including the man pages) and stripped as expected.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7247