Kill autotools
Summary:
This diff drops all the autotools-only files, removes the associated CI
build configuration and updates the `release-process.md` and
`release-notes.md` files.

There is still a little cleanup that can be done but will be part of
other diffs as this will change some code and I don't want to make it
more difficult to review.

Test Plan:
Git ls-filed and grepped for various patterns, the remaining are either
libs (secp256k1, univalue) or depends.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7244