txdb: don't reset during in-memory cache resize
Summary:
We can't support a reset of the dbwrapper object when in-memory configuration is used
because it results in the permanent loss of coins. This only affects unittest
configurations (since that's the only place we use in-memory CCoinsViewDB instances).

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19806 | core#19806]] [2/8]
https://github.com/bitcoin/bitcoin/pull/19806/commits/ad949ba449ff2115e3d22c71f5b6509f11112098

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11226