[CMAKE] Don't enforce LC_ALL=C.UTF-8 for the cmake test wrapper script
Summary:
The `LC_ALL=C.UTF-8` requirement is enforced to avoid locale causing
variance in shell scripts. But this setting is invalid on some linux
variants as it is not standard, and is causing more trouble than it is
useful.
Since the cmake util `test_wrapper.sh` is just a wrapper for a test
binary call which itself resets the locale, it can be simply discarded.
This will suppress a ton of warnings when running `make check` on
systems that don't support the `C.UTF-8` locale (e.g. Archlinux).
Furthermore it will make the result consistent with a direct call to
`test_bitcoin` for example.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7254