[CI] Prevent stdout buffer overflow
Summary:
This diff increases the `StreamReader` buffer size which is responsible
for handling the stdout and stderr output of the running process. It
also handles the exception in case of overflow so that the script will
continue to run and behave normally, with a notice that a line get
dropped.
This case currently occurs on CI during coverage builds:
https://build.bitcoinabc.org/viewLog.html?buildId=44175&buildTypeId=BitcoinABC_Master_BitcoinAbcMasterCoverage&tab=buildLog

Test Plan: Change the buffer size to something absurdly low and run a build.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7257