Add a script to update the AUR packages
Summary:
For each AUR package, this script:
 1. Completes/generates the package files
 2. Applies the changes to the upstream git repository
 3. Runs a full package build from the updated files
 4. Commit the changes and push to upstream

It requires an archlinux environment to run (with all the tools
installed) and a git user with write permission to the upstream
directory.

The sha256 checksums are removed from the `PKGBUILD` file as they are
regenerated by the script to reduce maintenance on the file.

The `python` package was missing as a dependency and would cause the
build to fail if not already installed on the user machine.

A README is added to explain how to update the packages using the
script.

Depends on D7260.

Test Plan:
Ran the build in an archlinux container and checked everything worked
excepted the final push.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7261